Pb21-1
Main struggles :  Making sure that all the neighbours were checked before checking the neighbours of neighbours.  So I used two "waiting lists". Also, I had a hard time handling the counter for the result, to not mix up the counts between the two waiting lists. I used generative AI only for the "boring tasks" at the beginning : reading file and saving it in a char, and then the basic function for the Vector class.